64 bits timestamps

+++ collector/BPFCollector.c
@@ -284,8 +284,8 @@ struct flow_info_t {
     u32 hop_latencies[MAX_INT_HOP];
     u16 queue_ids[MAX_INT_HOP];
     u16 queue_occups[MAX_INT_HOP];
-    u32 ingr_times[MAX_INT_HOP];
-    u32 egr_times[MAX_INT_HOP];
+    u64 ingr_times[MAX_INT_HOP];
+    u64 egr_times[MAX_INT_HOP];
     // u16 queue_congests[MAX_INT_HOP];
     u32 lv2_in_e_port_ids[MAX_INT_HOP];
     u32 tx_utilizes[MAX_INT_HOP];
